Create MIT LICENSE, Copyright 2021 Raku community
For choice of MIT license, see poll on issue #6.

For "Raku community", I copied what was in use by the most recent other Raku/* project that doesn't use "The Perl Foundation" or "The Free Software Foundation" (neither of which seems correct here). Unfortunately, that is from a little-known project called github.com/Raku/setup-raku/. Fortunately, it's by Shoichi Kaji, who has a fair amount of experience w/ project management. That being said, if there is any objection it's fine by me to modify it or request a change.