Remove g:raku_unicode_abbrevs feature
It's best to have the end user add these manually, rather than maintain
them in the plugin.

The documentation for them remains useful, I think.

I've modified the documented definition for × and ÷ to require a
trailing '+' character, which I believe should make them less obtrusive
but still accessible. Of course, a user can use whatever they like if we
are not defining them ourselves.

Closes #20